Move loader size tests to different file
