+++ test/slim_build_test.js
@@ -6,7 +6,6 @@ var testHelpers = require("./helpers");
 var optimize = require("../index").optimize;
 var intersection = require("lodash/intersection");
 var escapeRegExp = require("lodash/escapeRegExp");
-var checkSizeSnapshot = require("./check_size_snapshot");
 
 var fs = require("fs-extra");
 var rimraf = require("rimraf");
@@ -348,62 +347,6 @@ describe("slim builds", function() {
 				assert.ok(data[1], "circularWork should be true");
 				data[0](); // close();
 			});
-	});
-
-	describe("loader size benchmarks", function() {
-		it("single bundle", function() {
-			var base = path.join(__dirname, "slim", "basics");
-			var config = { config: path.join(base, "stealconfig.js") };
-
-			return rmdir(path.join(base, "dist"))
-				.then(function() {
-					return optimize(config, { quiet: true });
-				})
-				.then(function() {
-					return checkSizeSnapshot(
-						path.join(base, "dist", "bundles", "main.js"),
-						base
-					);
-				});
-		});
-
-		it("progressively loaded bundles", function() {
-			var base = path.join(__dirname, "slim", "progressive");
-			var config = { config: path.join(base, "stealconfig.js") };
-
-			return rmdir(path.join(base, "dist"))
-				.then(function() {
-					return optimize(config, { quiet: true });
-				})
-				.then(function() {
-					return Promise.all([
-						checkSizeSnapshot(
-							path.join(base, "dist", "bundles", "main.js"),
-							base
-						),
-						checkSizeSnapshot(
-							path.join(base, "dist", "bundles", "baz.js"),
-							base
-						)
-					]);
-				});
-		});
-
-		it("using plugins", function() {
-			var base = path.join(__dirname, "slim", "plugins");
-			var config = { config: path.join(base, "package.json!npm") };
-
-			return rmdir(path.join(base, "dist"))
-				.then(function() {
-					return optimize(config, { quiet: true });
-				})
-				.then(function() {
-					return checkSizeSnapshot(
-						path.join(base, "dist", "bundles", "plugins", "main.js"),
-						base
-					);
-				});
-		});
 	});
 
 	it("ESM named imports work", function() {
